polarion test case cleanup
