+++ cfme/tests/services/test_service_catalogs.py
@@ -202,345 +202,3 @@ def test_advanced_search_registry_element(request, appliance):
     request.addfinalizer(view.search.close_advanced_search)
     view.search.advanced_search_form.search_exp_editor.registry_form_view.fill({'type': "Registry"})
     assert view.search.advanced_search_form.search_exp_editor.registry_form_view.type.is_displayed
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_email_should_be_sent_when_service_approval_is_set_to_manual():
-    """ Email should be sent when service approval is set to manual
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.8
-        tags: service
-    Bugzilla:
-        1380197
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_role_configured_with_the_option_only_user_or_group_owned_should_allow_to_access_to_se():
-    """ Role configured with the option "only user or group owned
-        should allow to access to service catalogs and items
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.9
-        tags: service
-    Bugzilla:
-        1554775
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_notification_banner_vm_provisioning_notification_and_service_request_should_be_in_syn():
-    """ Notification Banner - VM Provisioning Notification and Service Request should be in  Sync
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.7
-        tags: service
-    Bugzilla:
-        1389312
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_database_wildcard_should_work_and_be_included_in_the_query():
-    """ Database wildcard should work and be included in the query
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.10
-        tags: service
-    Bugzilla:
-        1581853
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_user_should_be_able_to_see_requests_irrespective_of_tags_assigned():
-    """ User should be able to see requests irrespective of tags assigned
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        initialEstimate: 1/4h
-        testtype: functional
-        startsin: 5.9
-        tags: service
-    Bugzilla:
-        1641012
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_changing_action_order_in_catalog_bundle_should_not_removes_resource():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        initialEstimate: 1/16h
-        testtype: functional
-        startsin: 5.8
-        tags: service
-    Bugzilla:
-        1615853
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_notification_banner_service_event_should_be_shown_in_notification_bell():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        initialEstimate: 1/4h
-        testtype: functional
-        startsin: 5.9
-        tags: service
-        testSteps:
-            1. OPS UI  and SSUI service requests should create an event in notification bell
-            2. Also check , Clear All and "MArk as read" in notification bell
-            3. Number of events shown in notification bell
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_reconfigure_service_fields_empty_after_deploying_service():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.8
-        tags: service
-    Bugzilla:
-        1580987
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_order_service_after_deleting_provider():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.8
-        tags: service
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_custom_image_on_item_bundle_crud():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/8h
-        startsin: 5.5
-        tags: service
-        testSteps:
-            1. Create a catalog item
-            2. Upload custom image
-            3. remove custom image
-            4. Create a catalog  bundle
-            5. Upload a custom image
-            6. Change custom image
-        expectedResults:
-            1.
-            2. No error seen
-            3.
-            4.
-            5. No error seen
-    Bugzilla:
-        1487056
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_catalog_item_changing_the_provider_template_after_filling_all_tabs():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.5
-        tags: service
-    Bugzilla:
-        1240443
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_retire_ansible_service_bundle():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.5
-        tags: service
-    Bugzilla:
-        1363897
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_remove_catalog_items_from_catalog_bundle_resource_list():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-    Bugzilla:
-        1639557
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_request_filter_on_request_page():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.9
-        tags: service
-    Bugzilla:
-        1498237
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_check_all_availability_zones_for_amazon_provider():
-    """ Check if all availability zones can be selected while creating catalog item.
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/8h
-        startsin: 5.5
-        tags: service
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_service_retirement_requests_shall_be_run_by_the_user():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/8h
-        startsin: 5.5
-        tags: service
-        testSteps:
-            1. Create a request from non-admin user. Request shall be run by user
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_edit_catalog_item_after_remove_resource_pool():
-    """ Create catalog item with a resource pool , Remove resource pool from
-        the provider and then edit catalog item.
-    Polarion:
-        assignee: nansari
-        casecomponent: Services
-        testtype: functional
-        initialEstimate: 1/8h
-        startsin: 5.5
-        tags: service
-        testSteps:
-            1. Create a catalog item
-            2. Select cluster and resource pool and Save
-            3. Remove resource pool from provider
-            4. Edit catalog
-        expectedResults:
-            1.
-            2.
-            3.
-            4. Validation message should be shown
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_create_generic_instance():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Automate
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.9
-        tags: service
-    Bugzilla:
-        1577395
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(3)
-def test_generic_object_details_displayed_from_a_service_do_not_include_associations():
-    """
-    Polarion:
-        assignee: nansari
-        casecomponent: Automate
-        testtype: functional
-        initialEstimate: 1/4h
-        startsin: 5.9
-        tags: service
-    Bugzilla:
-        1576828
-    """
-    pass
